Clearance update
Request for splitters to be updated to accommodate 2" diameter filter installed on each outlet of splitter. Remodeled parts to work with Solidworks 2018. Note that expiratory splitter as previously configured (and modeled here) will not allow for installation of one-way valve flapper. Likely that all splitters will require redesign to better fit new manifold design and/or IM manufacturing processes.