Next revision of PCB (#152)
* Next revision of PCB

Changed I2C pullup resistors to 3Kohm

General tidy up and fix of issues on the PCB. Tweaked some polygons to eliminate sharp corners, made some traces larger.

* Changes made as per PR152

Changes made as requested, I also updated the pdf of the schematic and the dru file.

One thing I have noticed is that generating the BOM automatically comes up with the incorrect resistor value of 0.3 Ohms. I suspect this is to do with the library that is being used.

* Create READ ME BEFORE CONTINUING.md

Added a cautionary text file to explain the issue with the automatically generated BOM and the incorrect resistor values being utilised.